minir cleanup changes

+++ firmware/robot2015/src-ctrl/modules/TaskSignals.hpp
@@ -3,4 +3,4 @@
 // This is where the signals for different threads are defined for use across
 // mutliple files
 
-static const uint32_t MAIN_TASK_CONTINUE = 0x01;
+static const uint8_t MAIN_TASK_CONTINUE = 0x01;
